pyproject.toml
[tool.poetry]
name = "obey-the-testing-goat"
version = "0.1.0"
description = "Working code for Harry Percival's Obey the Testing Goat"
authors = ["Patrick Godschalk <patrick@kernelpanics.nl>"]
license = "UNLICENSED"
readme = "README.md"
packages = [{include = "obey_the_testing_goat"}]

[tool.poetry.dependencies]
python = "^3.6"
django = "<1.12"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
black = "<22"
flake8 = "<6"
selenium = "<4"
fabric3 = "^1.14.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
